+++ gitmap/constants/constants.go
@@ -14,7 +14,7 @@ package constants
 
 // Version holds the current release version of gitmap.
 // This is overridden at compile time via -ldflags during CI/CD.
-var Version = "6.162.0" // Overridden at compile time via -ldflags
+var Version = "6.163.0" // Overridden at compile time via -ldflags
 
 // RepoPath is set at build time via -ldflags.
 var RepoPath = ""